style: ignoring cert-err33-c linting warnings
cert-err33-c
Warns on unused function return values. Many of the standard library functions return a value that indicates if the call was successful. Ignoring the returned value can cause unexpected behavior if an error has occurred. The following functions are checked:

This check is an alias of check bugprone-unused-return-value with a fixed set of functions.

The check corresponds to a part of CERT C Coding Standard rule ERR33-C. Detect and handle standard library errors. The list of checked functions is taken from the rule, with following exception:

The check can not differentiate if a function is called with NULL argument. Therefore the following functions are not checked: mblen, mbrlen, mbrtowc, mbtowc, wctomb, wctomb_s

https://clang.llvm.org/extra/clang-tidy/checks/cert/err33-c.html